// src/app/interfaces/personagem.ts
export interface Personagem {
    name: string;
    foto: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    planeta: string;
    homeworld: string;
    especie: string;
    films: string[];
    species: string[];
    vehicles: string[];
    starships: string[];
    created: Date;
    edited: Date;
    url: string;
}